+++ src/app.py
@@ -10,6 +10,8 @@ import praw
 import streamlit as st
 
 from config import settings
+from pipelines.stt import get_speech_to_text
+from pipelines.tts import get_text_to_speech
 from schemas import MediaFile, Speaker
 from utils.media.audio import concatenate_audio_files, cut_audio, get_audio_duration
 from utils.media.video import (
@@ -24,8 +26,6 @@ from utils.media.video import (
 )
 from utils.reddit.post import get_reddit_object
 from utils.reddit.screenshot import take_comment_screenshot, take_post_screenshot
-from utils.stt import get_speech_to_text
-from utils.tts import get_text_to_speech
 
 # Initialize session state
 if "selected_video" not in st.session_state:
